calculateReward working with elapsed time

+++ src/I.tokenmanager.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 interface ITokenManager {
     function setRewardTokenAddress(address _rewardTokenAddress) external;
-    function transferReward(address _to, uint _amount) external;
+    function transferReward(address _to, uint256 _amount) external;
 }
